修复了一些问题
