+++ SimpleFFmpegGUI.WPF/Panels/CodeArgumentsPanel.xaml
@@ -11,7 +11,6 @@
     AllowDrop="True"
     mc:Ignorable="d">
     <UserControl.Resources>
-        <local:Int2StringConverter x:Key="Int2StringConverter" />
         <Style
             BasedOn="{StaticResource DefaultComboBoxStyle}"
             TargetType="{x:Type ComboBox}">
