+++ SimpleFFmpegGUI.WPF/Panels/CodeArgumentsPanel.xaml.cs
@@ -7,17 +7,13 @@ using SimpleFFmpegGUI.Model;
 using SimpleFFmpegGUI.WPF.ViewModels;
 using SimpleFFmpegGUI.WPF.ViewModels;
 using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
-using System.Globalization;
 using System.IO;
-using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
 using System.Windows.Documents;
 using System.Windows.Input;
 using System.Windows.Media;
@@ -122,39 +118,6 @@ namespace SimpleFFmpegGUI.WPF.Panels
                 }
             }
 
-        }
-    }
-    public class Int2StringConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            if (parameter is string s)
-            {
-                Dictionary<int, string> dic = null;
-                try
-                {
-                    dic = s.Split(';').Select(p => p.Split(':')).ToDictionary(p => int.Parse(p[0]), p => p[1]);
-                }
-                catch (Exception ex)
-                {
-                    throw new Exception("无法解析参数", ex);
-                }
-                if (value is int i)
-                {
-                    if (!dic.ContainsKey(i))
-                    {
-                        throw new Exception("没有值" + i + "的转换目标");
-                    }
-                    return dic[i];
-                }
-                throw new ArgumentException("绑定值必须为整数");
-            }
-            throw new ArgumentException("参数必须为字符串");
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
         }
     }
 }
