+++ SimpleFFmpegGUI.WPF/Panels/PresetsPanel.xaml
@@ -12,7 +12,6 @@
     d:DesignWidth="800"
     mc:Ignorable="d">
     <UserControl.Resources>
-        <local:Int2StringConverter x:Key="Int2StringConverter" />
         <Style
             BasedOn="{StaticResource DefaultComboBoxStyle}"
             TargetType="{x:Type ComboBox}">
